Fix error on Py 3 with nonascii subrequest.
The subrequest would succeed, but the non-ascii would be ugly.
Added a test for this.
With a French example, we expect "Actualités" but get "ActualitÃ©s".
Or with a Greek example we expect "Αρχική Σελίδα" but get "Î\x91Ï\x81Ï\x87Î¹ÎºÎ® Î£ÎµÎ»Î¯Î´Î±".
(This may or may not be readable in this commit message...)

Fixes https://github.com/plone/Products.CMFPlone/issues/3068
and https://github.com/plone/plone.app.theming/issues/162

This partially undoes https://github.com/plone/plone.app.theming/pull/146
and is basically the same as this fix from 2011:
https://github.com/plone/plone.app.theming/commit/1298857e0ec8c4846807fc6b2504b686f5384918
That fix was still there, but only executed on Python 2.  Now also Python 3.