Avoid Server Side Request Forgery via lxml parser.
Taken over from [`PloneHotfix20210518`](https://plone.org/security/hotfix/20210518/server-side-request-forgery-via-lxml-parser).

Strangely, several tests in `test_transforms` fail, seemingly because we try to load a theme html from the tile system.
This may depend on the `absolutePrefix` setting.
Commenting out the new FailingFileSystemResolver, helps.  But that is one of the parts that brings extra security.

In practice with the hotfix, I have not seen it fail.  Also, I would expect a big cry out if this part was broken.
But these failures are definitely  suspect.